+++ idea-plugin/src/main/kotlin/com/github/hoshinofw/multiversion/idea_plugin/OverwriteDuplicateClassHighlightFilter.kt
@@ -1,6 +1,7 @@
 package com.github.hoshinofw.multiversion.idea_plugin
 
 import com.github.hoshinofw.multiversion.engine.PathUtil
+import com.github.hoshinofw.multiversion.engine.VersionUtil
 import com.intellij.codeInsight.daemon.impl.HighlightInfo
 import com.intellij.codeInsight.daemon.impl.HighlightInfoFilter
 import com.intellij.lang.annotation.HighlightSeverity
@@ -9,7 +10,7 @@ import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiJavaFile
 import com.intellij.psi.search.GlobalSearchScope
 
-private val VERSION_DIR_REGEX = Regex("""/([0-9]+\.[0-9]+(\.[0-9]+)?)/""")
+private val VERSION_DIR_REGEX = Regex("/(${VersionUtil.VERSION_PATTERN.pattern})/")
 private val PATCHED_SEGMENT = "/${PathUtil.PATCHED_SRC_DIR}/"
 
 class OverwriteDuplicateClassHighlightFilter : HighlightInfoFilter {